Add basic websocket support from client to server
The only thing happening right now is that the client connects to
the server.

To avoid git collisions, the websocket code is written in a separate
temporary file client/src/pd-websocket.lua.

Replace the contents of main.lua with this file to test out websockets.
The websocket libs needs to be built first before this works.